+++ LeaderAnalytics.AdaptiveClient/LeaderAnalytics.AdaptiveClient/BaseClientFactory.cs
@@ -17,7 +17,7 @@ namespace LeaderAnalytics.AdaptiveClient
 
         public IEndPointConfiguration CachedEndPoint
         {
-            // "Current" and "Cached" are not the same - do not set EndPointContext.CurentEndPoint here.
+            // "Current" and "Cached" are not the same -  do not set EndPointContext.CurentEndPoint here.
             get { return endPointCache.GetEndPoint(perimeter.API_Name); }
 
             protected set { endPointCache.SetEndPoint(perimeter.API_Name, value); }
@@ -43,16 +43,10 @@ namespace LeaderAnalytics.AdaptiveClient
             perimeter = epcFactory(typeof(T));
 
             if (perimeter == null)
-                Hurl($"A Perimeter object could not be resolved for type {typeof(T).Name}.  A possible reason for this error is that type {typeof(T).Name} was not registered with a specific API_Name.", null);
+                throw new Exception($"A Perimeter object could not be resolved for type {typeof(T).Name}.  A possible reason for this error is that type {typeof(T).Name} was not registered with a specific API_Name.", null);
 
             if(perimeter.EndPoints == null || ! perimeter.EndPoints.Any())
-                Hurl($"A Perimeter object with an API_Name of {perimeter.API_Name} was resolved for type {typeof(T).Name},  however the EndPointConfigurationItems collection is null or contains no entries.", null);
-        }
-
-        protected void Hurl(string message, Exception innerEx)
-        {
-            Exception ex = new Exception(message, innerEx);
-            throw ex;
+                throw new Exception($"A Perimeter object with an API_Name of {perimeter.API_Name} was resolved for type {typeof(T).Name},  however the EndPointConfigurationItems collection is null or contains no entries.", null);
         }
 
         protected void SetAvailableEndPoints(string[] overrideNames)
@@ -63,7 +57,7 @@ namespace LeaderAnalytics.AdaptiveClient
                 AvailableEndPoints = perimeter.EndPoints.OrderBy(x => x.Preference);
 
             if (!AvailableEndPoints.Any())
-                Hurl($"No EndPointConfigurationItems were available to use for {typeof(T).Name}.  A possible reason for this exception is that no EndPointConfigurations were found with names that match the list of override names.", null);
+                throw new Exception($"No EndPointConfigurationItems were available to use for {typeof(T).Name}.  A possible reason for this exception is that no EndPointConfigurations were found with names that match the list of override names.", null);
 
             // Make sure the current EndPoint, if any, exists in the registered EndPoint list after being filtered for overrides.
             if (CachedEndPoint != null && !AvailableEndPoints.Any(x => x.Name == CachedEndPoint.Name))