+++ LeaderAnalytics.AdaptiveClient/LeaderAnalytics.AdaptiveClient/ClientEvaluator.cs
@@ -51,7 +51,7 @@ namespace LeaderAnalytics.AdaptiveClient
                     if (exceptions == null)
                         exceptions = new List<Exception>(10);
 
-                    exceptions.Add(new Exception(errorMsg, ex));  // We log but don't throw unless we run out of endpoints
+                    exceptions.Add(new Exception(errorMsg, ex));  //  We log but don't throw unless we run out of endpoints
                     logger?.Invoke(errorMsg);
                 }
             }
@@ -63,7 +63,7 @@ namespace LeaderAnalytics.AdaptiveClient
                 if (exceptions != null)
                     aggEx = new AggregateException(exceptions);
 
-                Hurl($"A functional EndPointConfiguration could not be resolved for client of type {typeof(T).Name}.  See inner exception(s) for more detail.", aggEx);
+                throw new Exception($"A functional EndPointConfiguration could not be resolved for client of type {typeof(T).Name}.  See inner exception(s) for more detail.", aggEx);
             }
         }
 
@@ -115,7 +115,7 @@ namespace LeaderAnalytics.AdaptiveClient
                 if (exceptions != null)
                     aggEx = new AggregateException(exceptions);
 
-                Hurl($"A functional EndPointConfiguration could not be resolved for client of type {typeof(T).Name}.  See inner exception(s) for more detail.", aggEx);
+                throw new Exception($"A functional EndPointConfiguration could not be resolved for client of type {typeof(T).Name}.  See inner exception(s) for more detail.", aggEx);
             }
         }
     }