+++ LeaderAnalytics.AdaptiveClient/tests/CallTests.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using NUnit.Framework;
 using LeaderAnalytics.AdaptiveClient;
+using LeaderAnalytics.AdaptiveClient.Utilities;
 using Autofac;
 using Autofac.Core;
 using Moq;
@@ -19,11 +20,13 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
         [Test]
         public void Reslove_InProcessClient1_of_type_IDummyAPI1()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Returns(true);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(false);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> MSSQL_Validator_Mock = new Mock<IEndPointValidator>();
+            MSSQL_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            builder.RegisterInstance(MSSQL_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
+
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
@@ -31,15 +34,19 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.AreEqual("Application_SQL1", client1.CurrentEndPoint.Name);
             Assert.AreEqual("InProcessClient1", result);
         }
-
+        
         [Test]
         public void Reslove_InProcessClient3_of_type_IDummyAPI1()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(It.Is<string>(z => z == EndPoints.First(y => y.Name == "Application_MySQL1").ConnectionString))).Returns(true);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(false);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> MSSQL_Validator_Mock = new Mock<IEndPointValidator>();
+            MSSQL_Validator_Mock.Setup(x => x.IsInterfaceAlive(It.IsAny<IEndPointConfiguration>())).Returns(false);
+            Moq.Mock<IEndPointValidator> MySQL_Validator_Mock = new Mock<IEndPointValidator>();
+            MySQL_Validator_Mock.Setup(x => x.IsInterfaceAlive(It.Is<IEndPointConfiguration>(z => z == EndPoints.First(y => y.Name == "Application_MySQL1")))).Returns(true);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            builder.RegisterInstance(MSSQL_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(MySQL_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MySQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
@@ -48,15 +55,17 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.AreEqual("InProcessClient3", result);
         }
 
-
+        
         [Test]
-        public void Reslove_InProcessClient3_of_type_IDummyAPI1_when_EndPaoint_name_is_passed()
+        public void Reslove_InProcessClient3_of_type_IDummyAPI1_when_EndPoint_name_is_passed()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Returns(true);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(false);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> InProcess_Validator_Mock = new Mock<IEndPointValidator>();
+            InProcess_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MySQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
@@ -69,14 +78,17 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.AreEqual("InProcessClient3", result2);
         }
 
+
         [Test]
         public void Reslove_WebAPIClient_of_type_IDummyAPI1()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Returns(false);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(true);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> InProcess_Validator_Mock = new Mock<IEndPointValidator>();
+            InProcess_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MySQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
@@ -84,39 +96,45 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.AreEqual("Application_WebAPI1", client1.CurrentEndPoint.Name);
             Assert.AreEqual("WebAPIClient1", result);
         }
-
+        
+        
         [Test]
         public void Throws_when_resolving_unregistered_client()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Returns(false);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(true);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> InProcess_Validator_Mock = new Mock<IEndPointValidator>();
+            InProcess_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MySQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
+
             Assert.Throws<DependencyResolutionException>(() => container.Resolve<IAdaptiveClient<IDummy3>>());
         }
 
+        
         [Test]
         public void Uses_cached_endpoint_on_second_call()
         {
             int inProcessCalls = 0;
             int webAPICalls = 0;
 
-
-            // NetworkUtilities Mock
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Callback(() => inProcessCalls++).Returns(false);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Callback(() => webAPICalls++).Returns(true);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> InProcess_Validator_Mock = new Mock<IEndPointValidator>();
+            InProcess_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Callback(() => inProcessCalls++).Returns(false);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Callback(() => webAPICalls++).Returns(true);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MySQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
+
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             string result = client1.Call(x => x.GetString());
             Assert.AreEqual("Application_WebAPI1", client1.CurrentEndPoint.Name);
             Assert.AreEqual("WebAPIClient1", result);
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
             Assert.AreEqual(1, webAPICalls);
             
             // do it again and use the cached endpoint:
@@ -125,19 +143,23 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             string result2 = client2.Call(x => x.GetString());
             Assert.AreEqual("Application_WebAPI1", client2.CurrentEndPoint.Name);
             Assert.AreEqual("WebAPIClient1", result2);
-            Assert.AreEqual(3, inProcessCalls);   // We should not test the in process endpoint again - we go directly to the cached HTTP endpoint.
+            Assert.AreEqual(2, inProcessCalls);   // We should not test the in process endpoint again - we go directly to the cached HTTP endpoint.
             Assert.AreEqual(1, webAPICalls);
         }
 
+        
 
         [Test]
         public void Client_exception_is_propagated()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Returns(true);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(false);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> InProcess_Validator_Mock = new Mock<IEndPointValidator>();
+            InProcess_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MySQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
+
 
             Moq.Mock<IDummyAPI1> inProcessClientMock = new Mock<IDummyAPI1>();
             inProcessClientMock.Setup(x => x.GetString()).Throws(new Exception("InProcess Exception"));
@@ -151,6 +173,64 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
 
             Assert.AreEqual("Application_SQL1", client1.CurrentEndPoint.Name);
             Assert.AreEqual("InProcess Exception", ex.Message);
+        }
+
+
+        
+        [Test]
+        public async Task CurrentEndPoint_is_thread_safe()
+        {
+            Moq.Mock<IEndPointValidator> InProcess_Validator_Mock = new Mock<IEndPointValidator>();
+            InProcess_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            builder.RegisterInstance(InProcess_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            IContainer container = builder.Build();
+
+            IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
+            IAdaptiveClient<IDummyAPI2> client2 = container.Resolve<IAdaptiveClient<IDummyAPI2>>();
+            string result1 = client1.Call(x => x.GetString());
+            Assert.AreEqual("Application_SQL1", client1.CurrentEndPoint.Name);
+            Assert.AreEqual("InProcessClient1", result1);
+            int result2 = client2.Call(x => x.GetInt());
+            Assert.AreEqual("Application_SQL2", client2.CurrentEndPoint.Name);
+            Assert.AreEqual(1, result2);
+
+            Task t1 = Task.Run(() => {
+
+                for (int i = 0; i < 1000; i++)
+                {
+                    string r1 = client1.Call(x => x.GetString());
+                    Assert.AreEqual("Application_SQL1", client1.CurrentEndPoint.Name);
+                    Assert.AreEqual("InProcessClient1", r1);
+
+                    int r2 = client2.Call(x => x.GetInt());
+                    Assert.AreEqual("Application_SQL2", client2.CurrentEndPoint.Name);
+                    Assert.AreEqual(1, r2);
+                }
+            });
+
+            Task t2 = Task.Run(() => {
+
+                for (int i = 0; i < 1000; i++)
+                {
+                    int r2 = client2.Call(x => x.GetInt());
+                    Assert.AreEqual("Application_SQL2", client2.CurrentEndPoint.Name);
+                    Assert.AreEqual(1, r2);
+
+                    string r1 = client1.Call(x => x.GetString());
+                    Assert.AreEqual("Application_SQL1", client1.CurrentEndPoint.Name);
+                    Assert.AreEqual("InProcessClient1", r1);
+                }
+            });
+
+            try
+            {
+                await Task.WhenAll(t1, t2);
+            }
+            catch (Exception)
+            {
+                throw;
+            }
+            
         }
     }
 }