+++ LeaderAnalytics.AdaptiveClient/tests/TryTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using LeaderAnalytics.AdaptiveClient;
+using LeaderAnalytics.AdaptiveClient.Utilities;
 using NUnit.Framework;
 using Autofac;
 using Moq;
@@ -15,11 +16,12 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
         [Test]
         public void Reslove_InProcessClient_of_type_IDummyAPI1()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Returns(true);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(false);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> MSSQL_Validator_Mock = new Mock<IEndPointValidator>();
+            MSSQL_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            builder.RegisterInstance(MSSQL_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
@@ -27,16 +29,17 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.AreEqual("Application_SQL1", client1.CurrentEndPoint.Name);
             Assert.AreEqual("InProcessClient1", result);
         }
-
+        
 
         [Test]
         public void Reslove_InProcessClientAsync_of_type_IDummyAPI1()
         {
-            Moq.Mock<INetworkUtilities> networkUtilMock = new Mock<INetworkUtilities>();
-            networkUtilMock.Setup(x => x.VerifyDBServerConnectivity(Moq.It.IsAny<string>())).Returns(true);
-            networkUtilMock.Setup(x => x.VerifyHttpServerAvailability(Moq.It.IsAny<string>())).Returns(false);
-            INetworkUtilities networkUtil = networkUtilMock.Object;
-            builder.RegisterInstance(networkUtil).As<INetworkUtilities>();
+            Moq.Mock<IEndPointValidator> MSSQL_Validator_Mock = new Mock<IEndPointValidator>();
+            MSSQL_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(true);
+            Moq.Mock<IEndPointValidator> HTTP_Validator_Mock = new Mock<IEndPointValidator>();
+            HTTP_Validator_Mock.Setup(x => x.IsInterfaceAlive(Moq.It.IsAny<IEndPointConfiguration>())).Returns(false);
+            builder.RegisterInstance(MSSQL_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.InProcess + ProviderName.MSSQL);
+            builder.RegisterInstance(HTTP_Validator_Mock.Object).Keyed<IEndPointValidator>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
@@ -44,8 +47,7 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.AreEqual("Application_SQL1", client1.CurrentEndPoint.Name);
             Assert.AreEqual("InProcessClient1", result);
         }
-
-
+        
         [Test]
         public void Reslove_WebAPIClient_of_type_IDummyAPI1()
         {
@@ -63,7 +65,7 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.AreEqual("WebAPIClient1", result);
         }
 
-
+        
         [Test]
         public void Uses_cached_endpoint_on_second_call()
         {
@@ -74,24 +76,22 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             IDummyAPI1 inProcessClient = inProcessClientMock.Object;
 
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MSSQL);
-            builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess);
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MySQL);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             string result = client1.Try(x => x.GetString());
             Assert.AreEqual("WebAPIClient1", result);
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
 
             // do it again and use the cached endpoint:
 
             IAdaptiveClient<IDummyAPI1> client2 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             string result2 = client2.Try(x => x.GetString());
             Assert.AreEqual("WebAPIClient1", result);
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
         }
-
-
+        
         [Test]
         public void Uses_cached_endpoint_on_third_call_after_override_on_second_call()
         {
@@ -102,14 +102,13 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             IDummyAPI1 inProcessClient = inProcessClientMock.Object;
 
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MSSQL);
-            builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess);
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MySQL);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             string result = client1.Try(x => x.GetString());
             Assert.AreEqual("WebAPIClient1", result);
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
 
             //  use the Application_WebAPI1A EndPoint by passing it to the client as an override
 
@@ -125,10 +124,10 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             string result2 = client2.Try(x => x.GetString());
             Assert.AreEqual("WebAPIClient1", result);
             Assert.AreEqual(client1.CurrentEndPoint.Name, "Application_WebAPI1A");
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
         }
 
-
+        
         [Test]
         public void Uses_next_endpoint_when_cached_endpoint_fails_on_second_call()
         {
@@ -139,20 +138,19 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             inProcessClientMock.Setup(x => x.GetString()).Callback(() => inProcessCalls++).Throws(new Exception("OMG"));
             IDummyAPI1 inProcessClient = inProcessClientMock.Object;
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MSSQL);
-            builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess);
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MySQL);
 
             Moq.Mock<IDummyAPI1> webAPIClientMock = new Mock<IDummyAPI1>();
             webAPIClientMock.Setup(x => x.GetString()).Callback(() => { if(webAPICalls++ == 1) throw new Exception("OMG"); }).Returns("WebAPIClient1");
             IDummyAPI1 webAPIClient = webAPIClientMock.Object;
-            builder.RegisterInstance(webAPIClient).Keyed<IDummyAPI1>(EndPointType.HTTP);
+            builder.RegisterInstance(webAPIClient).Keyed<IDummyAPI1>(EndPointType.HTTP + ProviderName.HTTP);
 
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             string result = client1.Try(x => x.GetString());
             Assert.AreEqual("WebAPIClient1", result);
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
             Assert.AreEqual(1, webAPICalls);
 
             // Application_WebAPI1 is the cached EndPoint.  We expect it to fail on this call and we use Application_WebAPI1A
@@ -168,10 +166,10 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             string result2 = client2.Try(x => x.GetString());
             Assert.AreEqual("WebAPIClient1", result);
             Assert.AreEqual(client1.CurrentEndPoint.Name, "Application_WebAPI1A");
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
         }
 
-
+                  
         [Test]
         public void Throws_when_invalid_EndPoint_name_is_passed_as_override()
         {
@@ -182,14 +180,13 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             IDummyAPI1 inProcessClient = inProcessClientMock.Object;
 
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MSSQL);
-            builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess);
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MySQL);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             string result = client1.Try(x => x.GetString());
             Assert.AreEqual("WebAPIClient1", result);
-            Assert.AreEqual(3, inProcessCalls);
+            Assert.AreEqual(2, inProcessCalls);
 
             // cached EndPoint is set but we are going to pass an invalid name
 
@@ -197,7 +194,7 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             Assert.Throws<Exception>(() => client2.Try(x => x.GetString(), "does not exist"));
         }
 
-
+        
         [Test]
         public void Inner_exceptions_are_maintained_for_each_client_call()
         {
@@ -209,28 +206,26 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             inProcessClientMock.Setup(x => x.GetString()).Throws(new Exception("InProcess Exception"));
             IDummyAPI1 inProcessClient = inProcessClientMock.Object;
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MSSQL);
-            builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess);
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MySQL);
 
             Moq.Mock<IDummyAPI1> webAPIClientMock = new Mock<IDummyAPI1>();
             webAPIClientMock.Setup(x => x.GetString()).Throws(new Exception("WebAPI Exception"));
             IDummyAPI1 webAPIClient = webAPIClientMock.Object;
-            builder.RegisterInstance(webAPIClient).Keyed<IDummyAPI1>(EndPointType.HTTP);
+            builder.RegisterInstance(webAPIClient).Keyed<IDummyAPI1>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             Exception ex = Assert.Throws<Exception>(() => client1.Try(x => x.GetString()));
-            Assert.AreEqual(5, ((AggregateException)(ex.InnerException)).InnerExceptions.Count);
+            Assert.AreEqual(4, ((AggregateException)(ex.InnerException)).InnerExceptions.Count);
             
             // Assert that the error thrown by the client is maintained
             Assert.AreEqual("InProcess Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[0].InnerException.Message);
             Assert.AreEqual("InProcess Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[1].InnerException.Message);
-            Assert.AreEqual("InProcess Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[2].InnerException.Message);
+            Assert.AreEqual("WebAPI Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[2].InnerException.Message);
             Assert.AreEqual("WebAPI Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[3].InnerException.Message);
-            Assert.AreEqual("WebAPI Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[4].InnerException.Message);
         }
 
-
+        
         [Test]
         public async Task Inner_exceptions_are_maintained_for_each_client_call_async()
         {
@@ -242,26 +237,24 @@ namespace LeaderAnalytics.AdaptiveClient.Tests
             inProcessClientMock.Setup(x => x.GetString()).Throws(new Exception("InProcess Exception"));
             IDummyAPI1 inProcessClient = inProcessClientMock.Object;
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MSSQL);
-            builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess);
             builder.RegisterInstance(inProcessClient).Keyed<IDummyAPI1>(EndPointType.InProcess + ProviderName.MySQL);
 
             Moq.Mock<IDummyAPI1> webAPIClientMock = new Mock<IDummyAPI1>();
             webAPIClientMock.Setup(x => x.GetString()).Throws(new Exception("WebAPI Exception"));
             IDummyAPI1 webAPIClient = webAPIClientMock.Object;
-            builder.RegisterInstance(webAPIClient).Keyed<IDummyAPI1>(EndPointType.HTTP);
+            builder.RegisterInstance(webAPIClient).Keyed<IDummyAPI1>(EndPointType.HTTP + ProviderName.HTTP);
             IContainer container = builder.Build();
 
             IAdaptiveClient<IDummyAPI1> client1 = container.Resolve<IAdaptiveClient<IDummyAPI1>>();
             Exception ex = Assert.ThrowsAsync<Exception>(async () => await client1.TryAsync(async x => { await Task.Delay(0); x.GetString(); }));
             
-            Assert.AreEqual(5, ((AggregateException)(ex.InnerException)).InnerExceptions.Count);
+            Assert.AreEqual(4, ((AggregateException)(ex.InnerException)).InnerExceptions.Count);
 
             // Assert that the error thrown by the client is maintained
             Assert.AreEqual("InProcess Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[0].InnerException.Message);
             Assert.AreEqual("InProcess Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[1].InnerException.Message);
-            Assert.AreEqual("InProcess Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[2].InnerException.Message);
+            Assert.AreEqual("WebAPI Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[2].InnerException.Message);
             Assert.AreEqual("WebAPI Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[3].InnerException.Message);
-            Assert.AreEqual("WebAPI Exception", ((AggregateException)(ex.InnerException)).InnerExceptions[4].InnerException.Message);
         }
     }
 }